C++ homework --pointer practice
This is a single file that I think is currently broken,  it will become the basis of what i will create.   I am making now a demonstration of the pointer within a custom function that I want to be able to test seperately.   --Then I will think of a way to integrate it into the master payroll program.   thank you if you wish to offer assistance or comment.